Move and improve artist controls
