+++ lib/admin/photos/artists.module.css
@@ -1,0 +1,41 @@
+.Artists {
+    padding: 10px;
+    margin-top: 10px;
+    border: 1px solid gray;
+    border-radius: 10px;
+}
+
+.Artists h3 {
+    margin-bottom: 4px;
+    margin-top: 0px;
+}
+
+.Artists select {
+    margin-right: 10px;
+}
+
+.Artists button,
+.Artists select {
+    font-size: inherit;
+    margin-top: 4px;
+    padding: 10px;
+    cursor: pointer;
+    border: 1px solid gray;
+    border-radius: 10px;
+    background-color: transparent;
+    transition: border-color 0.1s ease-in-out;
+}
+
+.Artists button:hover,
+.Artists select:hover {
+    border-color: black;
+}
+
+.Entry {
+    display: flex;
+    align-items: center;
+}
+
+.Entry div {
+    flex-grow: 1;
+}
